ASas

+++ tasks/associative_operation/associative_operation.cpp
@@ -1,10 +1,10 @@
 #include "associative_operation.h"
 
 bool IsAssociative(const std::vector<std::vector<size_t>>& table) {
-    for (int x = 0; x<table.size(); ++x){
-        for (int y = 0; y<table.size(); ++y){
-            for (int z = 0; z<table.size(); ++z){
-                if (table[table[x][y]][z] != table[x][table[y][z]]){
+    for (size_t x = 0; x < table.size(); ++x) {
+        for (size_t y = 0; y < table.size(); ++y) {
+            for (size_t z = 0; z < table.size(); ++z) {
+                if (table[table[x][y]][z] != table[x][table[y][z]]) {
                     return false;
                 }
             }
